Refactor pages to components
