+++ static/js/components/Drawer_test.js
@@ -6,7 +6,7 @@ import { Drawer as RMWCDrawer } from "@rmwc/drawer"
 import Drawer from "./Drawer"
 
 import IntegrationTestHelper from "../util/integration_test_helper"
-import { toggleDrawer } from "../reducers/drawer"
+import { setDrawerOpen } from "../reducers/drawer"
 
 describe("Drawer", () => {
   let helper, render
@@ -25,20 +25,20 @@ describe("Drawer", () => {
       {
         children: <div className="testtesttest">test</div>
       },
-      [toggleDrawer()]
+      [setDrawerOpen(true)]
     )
     assert.ok(wrapper.find(RMWCDrawer).exists())
     assert.equal(wrapper.find(".testtesttest").text(), "test")
   })
 
   it("should have the drawer open if action is dispatch", async () => {
-    const { wrapper } = await render({}, [toggleDrawer()])
+    const { wrapper } = await render({}, [setDrawerOpen(true)])
     assert.isTrue(wrapper.find(RMWCDrawer).prop("open"))
   })
 
   it("should pass down a close function to drawer", async () => {
-    const { wrapper, store } = await render({}, [toggleDrawer()])
+    const { wrapper, store } = await render({}, [setDrawerOpen(false)])
     wrapper.find(RMWCDrawer).prop("onClose")()
-    assert.isFalse(store.getState().drawer.showDrawer)
+    assert.isFalse(store.getState().drawer.drawerOpen)
   })
 })
