+++ static/js/components/EditProfileDisplay.js
@@ -8,21 +8,20 @@ import { mutateAsync } from "redux-query"
 import { createStructuredSelector } from "reselect"
 import { MetaTags } from "react-meta-tags"
 
-import { EDIT_PROFILE_PAGE_TITLE } from "../../constants"
-import users, { currentUserSelector } from "../../lib/queries/users"
-import { routes } from "../../lib/urls"
-import queries from "../../lib/queries"
-import { formatTitle } from "../../util/util"
+import { EDIT_PROFILE_PAGE_TITLE } from "../constants"
+import users, { currentUserSelector } from "../lib/queries/users"
+import queries from "../lib/queries"
+import { formatTitle } from "../util/util"
 
-import EditProfileForm from "../../components/forms/EditProfileForm"
+import EditProfileForm from "./forms/EditProfileForm"
 
-import type { RouterHistory } from "react-router"
 import type {
   Country,
   CurrentUser,
   User,
   HttpAuthResponse
-} from "../../flow/authTypes"
+} from "../flow/authTypes"
+import { setDrawerState } from "../reducers/drawer"
 
 type StateProps = {|
   countries: ?Array<Country>,
@@ -30,23 +29,19 @@ type StateProps = {|
 |}
 
 type DispatchProps = {|
-  editProfile: (userProfileData: User) => Promise<HttpAuthResponse<User>>
-|}
-
-type ProfileProps = {|
-  history: RouterHistory
+  editProfile: (userProfileData: User) => Promise<HttpAuthResponse<User>>,
+  updateDrawer: (state: Object) => void
 |}
 
 type Props = {|
   ...StateProps,
-  ...DispatchProps,
-  ...ProfileProps
+  ...DispatchProps
 |}
 
-export class EditProfilePage extends React.Component<Props> {
+export class EditProfileDisplay extends React.Component<Props> {
   async onSubmit(profileData: User, { setSubmitting, setErrors }: Object) {
-    const { editProfile, history } = this.props
-
+    const { editProfile, updateDrawer } = this.props
+    console.log("In onSubmit")
     const payload = {
       ...profileData,
       ...(profileData.profile ?
@@ -57,7 +52,6 @@ export class EditProfilePage extends React.Component<Props> {
         } :
         {})
     }
-    console.log(payload)
     try {
       const {
         body: { errors }
@@ -65,16 +59,14 @@ export class EditProfilePage extends React.Component<Props> {
       { body: Object } = await editProfile(payload)
 
       if (errors && errors.length > 0) {
-        console.log(errors)
         setErrors({
           email: errors[0]
         })
       } else {
-        history.push(routes.profile.view)
+        updateDrawer("profileView")
       }
     } finally {
       setSubmitting(false)
-      console.log("done")
     }
   }
 
@@ -103,9 +95,10 @@ const mapStateToProps = createStructuredSelector({
   countries:   queries.users.countriesSelector
 })
 
-const mapDispatchToProps = {
-  editProfile: editProfile
-}
+const mapDispatchToProps = dispatch => ({
+  editProfile:  (data: User) => dispatch(editProfile(data)),
+  updateDrawer: (newState: ?string) => dispatch(setDrawerState(newState))
+})
 
 const mapPropsToConfigs = () => [
   queries.users.countriesQuery(),
@@ -115,4 +108,4 @@ const mapPropsToConfigs = () => [
 export default compose(
   connect(mapStateToProps, mapDispatchToProps),
   connectRequest(mapPropsToConfigs)
-)(EditProfilePage)
+)(EditProfileDisplay)
