+++ static/js/pages/DrawerPage.js
@@ -1,5 +1,5 @@
 // @flow
-import React, {useCallback} from "react"
+import React from "react"
 import { connect } from "react-redux"
 import { useDispatch, useSelector } from "react-redux"
 
@@ -7,50 +7,57 @@ import Drawer from "../components/Drawer"
 
 import { drawerSelector } from "../lib/selectors"
 import { setDrawerOpen, setDrawerState } from "../reducers/drawer"
-import EditProfilePage from "./profile/EditProfilePage";
-import ViewProfilePage from "./profile/ViewProfilePage";
+import EditProfileDisplay from "../components/EditProfileDisplay"
+import ViewProfileDisplay from "../components/ViewProfileDisplay"
 
-type Props = {
-}
+type Props = {}
 
+// $FlowFixMe: this is a temporary page
 export function DrawerPage(props: Props) {
-  const { drawerOpen, drawerState } = useSelector(drawerSelector)
+  const { setDrawerState } = props
+  const { drawerState } = useSelector(drawerSelector)
   const dispatch = useDispatch()
-  const closeDrawer = () =>
-    dispatch(setDrawerOpen(false))
+  const closeDrawer = () => dispatch(setDrawerOpen(false))
 
-  const openDrawer = (state) => {
+  const openDrawer = state => {
     dispatch(setDrawerOpen(true))
     dispatch(setDrawerState(state))
   }
 
-
   const renderChild = () => {
     switch (drawerState) {
-          case "profileEdit":
-            return <EditProfilePage />
-          case "profileView":
-            return <ViewProfilePage />
+    case "profileEdit":
+      return <EditProfileDisplay />
+    case "profileView":
+      return <ViewProfileDisplay />
     }
   }
 
-  return <>
-    <Drawer className="align-right">
-      {
-        renderChild()
-      }
-    </Drawer>
-    <ul>
-      <li><button onClick={() => closeDrawer()}>Close Drawer</button></li>
-      <li><button onClick={() => openDrawer("profileEdit")}>Edit Profile</button></li>
-      <li><button onClick={() => openDrawer("profileView")}>View Profile</button></li>
-    </ul>
-  </>
+  return (
+    <>
+      <Drawer className="align-right">{renderChild()}</Drawer>
+      <ul>
+        <li>
+          <button onClick={() => closeDrawer()}>Close Drawer</button>
+        </li>
+        <li>
+          <button onClick={() => openDrawer("profileEdit")}>
+            Edit Profile
+          </button>
+        </li>
+        <li>
+          <button onClick={() => openDrawer("profileView")}>
+            View Profile
+          </button>
+        </li>
+      </ul>
+    </>
+  )
 }
 
 const mapStateToProps = state => ({
   drawerState: state.drawer.drawerState,
-  drawerOpen:  state.drawer.drawerOpen,
+  drawerOpen:  state.drawer.drawerOpen
 })
 
 const mapDispatchToProps = dispatch => ({
